pyproject.toml
[project]
name = "cox_mate"
version = "0.1.0"
description = "A fun project to track your chambers points and drops"
authors = [
    {name = "Matthew Mulhall",email = "mattlm0831@gmail.com"}
]
license = {text = "Apache 2.0"}
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "transformers>=4.37.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=8.3.0",
    "accelerate>=0.20.0",
    "sentencepiece>=0.1.99",
    "sqlalchemy>=2.0.0",
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "dspy-ai[google] (>=3.1.3,<4.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
]

[project.optional-dependencies]
dspy = ["dspy-ai>=2.4.0"]
dev = ["pytest>=7.0.0", "pytest-mock>=3.10.0", "black", "isort", "flake8"]

[tool.poetry]
packages = [{include = "cox_mate"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
